+++ docs/changes.md
@@ -3,6 +3,7 @@ Release date: `2017-??-??`
 
 ### Core
 - [NXDRIVE-1059](https://jira.nuxeo.com/browse/NXDRIVE-1059): Wrong URL for the beta channel
+- [NXDRIVE-1061](https://jira.nuxeo.com/browse/NXDRIVE-1061): Remote rename of an accentued folder on Windows fails
 
 
 # 3.0.1